+++ apps/backend/src/libs/modules/server-application/server-application.ts
@@ -12,6 +12,8 @@ import { courseSectionController } from "~/modules/course-sections/course-sectio
 import { courseController } from "~/modules/courses/courses.js";
 import { fileController } from "~/modules/files/files.js";
 import { friendController } from "~/modules/friends/friends.js";
+import { groupController } from "~/modules/groups/groups.js";
+import { permissionController } from "~/modules/permissions/permissions.js";
 import { sectionStatusController } from "~/modules/section-statuses/section-statuses.js";
 import { userCourseController } from "~/modules/user-courses/user-courses.js";
 import { userNotificationController } from "~/modules/user-notifications/user-notifications.js";
@@ -33,6 +35,8 @@ const apiV1 = new BaseServerApplicationApi(
 	...userCourseController.routes,
 	...fileController.routes,
 	...friendController.routes,
+	...groupController.routes,
+	...permissionController.routes,
 	...userController.routes,
 	...userNotificationController.routes,
 	...courseSectionController.routes,