+++ apps/frontend/src/libs/constants/constants.ts
@@ -2,4 +2,8 @@ export { BACK_NAVIGATION_STEP } from "./back-navigation-step.constants.js";
 export { MENU_ITEMS } from "./menu-items.constants.js";
 export { PAGES_WITH_SEARCH_BAR } from "./pages-with-search-bar.constant.js";
 export { PAGINATION_PAGES_CUT_COUNT } from "./pagination-pages-cut-count.constant.js";
-export { EMPTY_ARRAY_LENGTH } from "@trackmates/shared";
+export {
+	EMPTY_ARRAY_LENGTH,
+	LAST_ARRAY_ITEM,
+	PREVIOUS_INDEX_OFFSET,
+} from "@trackmates/shared";