+++ apps/frontend/src/pages/overview/overview.tsx
@@ -1,5 +1,13 @@
-import { Courses, Loader, Pagination } from "~/libs/components/components.js";
-import { PAGINATION_PAGES_CUT_COUNT } from "~/libs/constants/constants.js";
+import {
+	Courses,
+	EmptyPagePlaceholder,
+	Loader,
+	Pagination,
+} from "~/libs/components/components.js";
+import {
+	EMPTY_ARRAY_LENGTH,
+	PAGINATION_PAGES_CUT_COUNT,
+} from "~/libs/constants/constants.js";
 import { DataStatus, PaginationValue } from "~/libs/enums/enums.js";
 import {
 	useAppDispatch,
@@ -52,26 +60,34 @@ const Overview: React.FC = () => {
 				userId: user.id,
 			}),
 		);
-	}, [dispatch, user, page]);
+	}, [dispatch, user, page, totalCount]);
+
+	const hasCourses = courses.length > EMPTY_ARRAY_LENGTH;
 
 	return (
 		<div className={styles["container"]}>
 			<WelcomeHeader onAddCourseClick={handleModalOpen} user={user} />
 			<div className={styles["courses-container"]}>
-				<div className={styles["courses-container-content"]}>
-					<h2 className={styles["courses-title"]}>Courses</h2>
-					{isLoading ? (
-						<Loader color="orange" size="large" />
-					) : (
-						<Courses courses={courses} userId={user.id} />
-					)}
-				</div>
-				<Pagination
-					currentPage={page}
-					onPageChange={handlePageChange}
-					pages={pages}
-					pagesCount={pagesCount}
-				/>
+				<h2 className={styles["courses-title"]}>Courses</h2>
+				{isLoading ? (
+					<Loader color="orange" size="large" />
+				) : (
+					<>
+						{hasCourses ? (
+							<div className={styles["courses-container-content"]}>
+								<Courses courses={courses} userId={user.id} />
+								<Pagination
+									currentPage={page}
+									onPageChange={handlePageChange}
+									pages={pages}
+									pagesCount={pagesCount}
+								/>
+							</div>
+						) : (
+							<EmptyPagePlaceholder title="You haven't added any courses yet" />
+						)}
+					</>
+				)}
 			</div>
 			{isAddCourseModalOpen && <AddCourseModal onClose={handleModalClose} />}
 		</div>